review fixes: + mv toggle favourite api view to api.py + mv sidebar_layout.html template to common/ folder
