+++ club/urls.py
@@ -18,9 +18,10 @@ from posts.models import Post
 from posts.rss import NewPostsRss
 from posts.sitemaps import sitemaps
 from posts.views.admin import admin_post, announce_post
+from posts.views.api import toggle_post_favourite
 from posts.views.feed import feed
 from posts.views.posts import show_post, edit_post, upvote_post, retract_post_vote, compose, compose_type, \
-    toggle_post_subscription, toggle_post_favourite, favourite_posts
+    toggle_post_subscription, favourite_posts
 from search.views import search
 from users.api import api_profile
 from users.views import profile, edit_profile, on_review, banned, rejected, intro, toggle_tag, \
