+++ frontend/html/posts/feed.html
@@ -1,4 +1,4 @@
-{% extends "sidebar_layout.html" %}
+{% extends "common/sidebar_layout.html" %}
 {% load text_filters %}
 {% load static %}
 {% load paginator %}
